Added CMake build system.
I added a CMake build system for the data files, which links in with the files from the main repository. It also has the capability to install the files along with the binary. Files added to the FOLDERNAME_FILES variable within each folder will be added to the build system, and will be copied to the build directory, and will also install along with the binary to the install prefix.