Added some basic configuration files for loader.
Added some files to complement the new loader, these files are just basic files, as I will probably be changing their structure as time goes on, but it gives me a base to start working on some gameplay and finding exactly which fields I will need for each file and how best to load them and everything. The filetypes are json, and are parsed by the jansson library.