Start adding real configurations that are parsed
I'm starting to write actual configurations that the game code can actually use
and parse, they will be slowly added, as I slowly add on new classes to the new
config module, which handles interfacing the textual json data files with the
rest of the code. There won't be much in here at the moment, just the bare
minimum to get it started and working. Currently this is just a file that is
used by the new ConfigGlobal class, and currently it just has the names of each
of the directories and the files that they will contain.