Fix crosscompiling with native lz4comp execution
Cross compiling builds were slightly broken due to the fact that during the build
process, lz4comp is called to compress the json files into lz4 files. This was broken
on cross compiling builds as it would try to execute the binary, as it would be in a
different executable format, and would cause data files not to be installed correctly.
This little bit of logic fixes that, it will take a cmake output file containing the
information for the native version, and will use that to compress the files. By
default, it will look for the file within a build sister directory, as when I cross
compile, I use a build directory, and for cross compiling, I use build-machine, such
as build-pandora. So it will automatically work without having to modify the information,
but it can be changed within ccmake or any of the gui utilities.