Support new json compression in lz4comp
A slight change to the build process, using the new capabilities of lz4comp so
that the maximum possible compression is achieved. Currently it isn't very
useful, only shaving 6 bytes off of the config.json file, but when I start
dealing with things like levels and arrays of data, the savings will be quite
substantial.